Sub-second precision for reveal times
DeviceRef.DelaySeconds is now a double (was int). The user can enter
"10.0", "10.1", "10.250" — fractional seconds parse correctly and round-
trip through the JSON. Old integer profile values still parse fine
(JSON numbers are double-typed).

Why this matters: the reveal sequence has to land every device inside
the game's hot-plug detection window. Some games stop accepting new HID
arrivals a few seconds into the session; with whole-second granularity
and a slow wheel driver, a six-device profile (10s wheel-warmup + 1s per
device) easily ran past the window. The user's MOZA + handbrake setup
was losing every device after handbrake.

With ms precision, "10.0 / 10.1 / 10.2 / 10.3 / 10.4 / 10.5" packs
six reveals into 500ms — comfortably inside any reasonable hot-plug
window.

LaunchOrchestrator.RevealDisableThenRestore now does all timing math in
milliseconds as doubles, with Task.Delay receiving an int ms argument.
Log format strings use {0:0.##} so "T+10s" stays "T+10s" but "T+10.5s"
displays cleanly when fractional.

UI: the TextBox uses StringFormat=0.### so trailing zeros are stripped
("5" instead of "5.000") but decimals are accepted as-is. Width bumped
32→40 to fit "10.5". UpdateSourceTrigger=LostFocus so the user can type
"10.5" without the binding fighting them mid-edit. Tooltip rewritten
with explicit guidance on packing reveals tight.

No schema bump — old integer "delaySeconds" values parse as doubles
transparently. Old profiles continue to work without migration.

+++ app/Models/Profile.cs
@@ -73,8 +73,13 @@ public class DeviceRef
     [JsonPropertyName("friendlyName")]
     public string FriendlyName { get; set; } = "";
 
+    /// <summary>
+    /// Absolute reveal time in seconds from start of the reveal phase, with
+    /// sub-second precision. Old integer values still parse (JSON numbers
+    /// are double-typed).
+    /// </summary>
     [JsonPropertyName("delaySeconds")]
-    public int DelaySeconds { get; set; } = 0;
+    public double DelaySeconds { get; set; } = 0;
 }
 
 public enum TriggerMode { HandleWatcher, Timer }
